Setting "Show Description" toggle to False in CPE doesn't save attribute to Flow Properties Fixes #18
See discussion on issue.
Flow Builder Runtime will send Strings instead
of Boolean values to CPE, so this update transforms
those strings to Booleans, and then back again.

Also changed the approach, using "hideDescription" with a default of False
instead of "showDescription" with a default
of True. Bad practice to use true as default
for Element attributes, so flipped that.

This is an "under the hood" change, as
CPE UI still presents as "Show Description"
because from the User perspective, it feels
more natural to be active when you are showing
the descriptions, and then toggle off to hide